ui: settings views
