+++ LDKNodeMonday/View Model/Profile/Peer/PeersListViewModel.swift
@@ -9,18 +9,10 @@ import LDKNode
 import SwiftUI
 
 class PeersListViewModel: ObservableObject {
-    @Published var networkColor = Color.gray
     @Published var peers: [PeerDetails] = []
 
     func listPeers() {
         self.peers = LightningNodeService.shared.listPeers()
-    }
-
-    func getColor() {
-        let color = LightningNodeService.shared.networkColor
-        DispatchQueue.main.async {
-            self.networkColor = color
-        }
     }
 
 }
