+++ LDKNodeMonday/View/Home/NetworkSettingsView.swift
@@ -53,7 +53,8 @@ struct NetworkSettingsView: View {
                 ToolbarItem(placement: .navigationBarTrailing) {
                     Button("Done") {
                         dismiss()
-                    }.padding()
+                    }
+                    .padding()
                 }
             }
         }
